refactoring: 입력 기능을 InputView로 클래스화 및 기능 축소
InputView 추가:
- main/kotlin/racingcar에 view 디렉토리 및 InputView.kt 추가
- 입력에 포함되었던 유효 입력 검사 기능을 축출하여 축소
